Removed SLAMDemo and added actual SLAM module
Added SLAM module and some additional functionality, such as
MapData class, which represents the map, and some minor utility
classes in the utilities header file. Also modified the J2B2Demo
concerning the SLAM visualization, but haven't tested this yet
in acion (it compiles ok). Removed the simple UI example from
the makefile, and added the new classes to it.